mbox, patches: add new review-mbox, review-patches scripts
The review-mbox script retrieves an mbox of all the emails b4 found related
to the series under review, saving it locally as a "review_[name].mbx" file.

This is useful for grepping or manual operations against the entire
contents of the thread.

The review-patches series goes one further and retrieves all the patches
for the specified series (excluding any cover letter) and saving them
locally as *.patch files.

A version can be specified, if none is then the latest version is
retrieved.

This is useful for manual operations against patches in the series.