mm_tests: fix errors in implementation, treat FAIL=0 as success
There were some mistakes due to changing implementation + not updating
implementation, we start by fixing these so run_mm_tests() is
paramter-less.

We correct the $logfile parameter not necessarily existing here.

We then handle the fact that run_vm_tests.sh will provide an exit status
even if no tests fails with a grep, so we don't fail other parts of scripts
if this incorrect fails.