start-applied: add new script to start review even if series applied
Sometimes a maintainer will update their tree with a series before you can
start reviewing it.

The review-start-applied command allows for setting up a review as normal
but using commits that already exist in the tree to account for this.

All other review tools should work as normal once this is done.