mm-tests: check locking, memleaks on mm tests, other improvements
Add a specific config kernel build and add a hook for it, and a command -
review-config-debug.

Update all the mm tests logic to use this config.

Importantly, it enables lockdep checks and kmemleak. So change the mm-tests
run to use a hook file which scans for kmemleaks right after the test run
and outputs any found.

Create a shared debug vng run function, vng_run_debug, and move the
existing config settings from review-vng to shared code.

Then use this debug-mode vng (as manually invoked by review-vng-debug) for
the mm-test runs.