config: enable -werror on debug builds, disable kasan on default
We disable CONFIG_WERROR for development purposes, but it's obviously
useful for debugging, so renable for debug configuration.

We may as well group all the debug stuff together, though we keep
CONFIG_DEBUG_VM enabled across default/debugging as it's useful to enable
VM_WARN_ON() etc. stuff and it's not too much overhead.