mk: add new review-mk command
Add a new simple build tool which rather hacks things to get the following
semantics:

* Build the kernel showing all stdout but...
* If an error arises, stop showing stdout and show only stedrr and...
* As soon as we see the 'unfinished jobs...' message, kill the build
  altogether.

This way you get the visual feedback of the usual output as well as not
losing error messages in noise.

NOTE: This patch was initially taken from LLM output to speed things up
(mea culpa), but then HEAVILY adapted and checked.