build-commits: fix variable scope and make shared function
We are accessing $from, $to without them being parameters to do_build() by
mistake.

Fix this and make do_per_commit_build() a shared function so we can use
elsewhere, also allow for parameters to be passed to the build in order to
make it easier in future to do cross-compile/clang builds this way.